requirements.txt
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
xgboost==1.7.5
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
streamlit==1.24.0
shap==0.42.1
pytest==7.3.1
black==23.3.0
isort==5.12.0
jupyter==1.0.0
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
